# miri: add some chance to reuse addresses of previously freed allocations