@rust-log-analyzer 2024-03-09
The job **`x86_64-gnu-tools`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/8215434662/job/22468897764) [(plain)](https://github.com/rust-lang/rust/commit/6dea495d2587480422b98f0ae34ad04e77c74fb1/checks/22468897764/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
GITHUB_ENV=/home/runner/work/_temp/_runner_file_commands/set_env_922ec5c3-fe00-4c15-99a2-9729fb28b264
GITHUB_EVENT_NAME=pull_request
GITHUB_EVENT_PATH=/home/runner/work/_temp/_github_workflow/event.json
GITHUB_GRAPHQL_URL=https://api.github.com/graphql
GITHUB_HEAD_REF=miri-addr-reuse
GITHUB_JOB=pr
GITHUB_PATH=/home/runner/work/_temp/_runner_file_commands/add_path_922ec5c3-fe00-4c15-99a2-9729fb28b264
GITHUB_REF=refs/pull/122240/merge
GITHUB_REF_NAME=122240/merge
GITHUB_REF_PROTECTED=false
---
   Compiler: "MIRI_ENV_VAR_TEST"="0" "MIRI_TEMP"="/tmp" /checkout/obj/build/x86_64-unknown-linux-gnu/stage2/bin/miri "--error-format=json" "-Dwarnings" "-Dunused" "-Ainternal_features" "-Zui-testing" "--target" "i686-pc-windows-msvc" "--out-dir" OUT_DIR
Error: 

   0: tests failed
FAILED TEST: tests/pass/address-reuse.rs
command: MIRI_ENV_VAR_TEST="0" MIRI_TEMP="/tmp" "/checkout/obj/build/x86_64-unknown-linux-gnu/stage2/bin/miri" "--error-format=json" "-Dwarnings" "-Dunused" "-Ainternal_features" "-Zui-testing" "--target" "i686-pc-windows-msvc" "--out-dir" "/checkout/obj/build/x86_64-unknown-linux-gnu/stage2-tools/ui/tests/pass" "tests/pass/address-reuse.rs" "--edition" "2021"

Backtrace omitted. Run with RUST_BACKTRACE=1 environment variable to display it.
Run with RUST_BACKTRACE=full to include source snippets.
error: pass test got exit status: 101, but expected 0

error: actual output differed from expected
error: actual output differed from expected
Execute `./miri test --bless` to update `tests/pass/address-reuse.stderr` to the actual output
--- tests/pass/address-reuse.stderr

Caused by:
  process didn't exit successfully: `/checkout/obj/build/x86_64-unknown-linux-gnu/stage2-tools/x86_64-unknown-linux-gnu/release/deps/compiletest-fa25a6b78de102f0 --quiet` (exit status: 1)
+++ <stderr output>
+++ <stderr output>
+thread 'main' panicked at $DIR/address-reuse.rs:LL:CC:
+assertion failed: addrs.len() < count


full stderr:
thread 'main' panicked at tests/pass/address-reuse.rs:12:5:
thread 'main' panicked at tests/pass/address-reuse.rs:12:5:
assertion failed: addrs.len() < count

full stdout:


```

</details>